Materials

CoupledDipole.Material — Type

Material(wavelength, media)

Particle cluster specification.

NOTE: for now all particles have the same material; easy to extend if needed, by making it a vector and matching a dictionary in HighLevel functions

Dielectric functions

CoupledDipole.epsilon_Ag — Function

epsilon_Ag(λ::Real)

Drude model for the dielectric function of silver in the visible region

λ: wavelength in nm

Examples

julia> round(epsilon_Ag(632.8), digits=5)
-16.11377 + 0.74871im

CoupledDipole.epsilon_Au — Function

epsilon_Au(λ::Real)

Extended Drude model for the dielectric function of gold in the visible region

λ: wavelength in nm

Examples

julia>  round(epsilon_Au(632.8), digits=5)
-11.40185 + 1.18679im

Particle polarisabilities

CoupledDipole.alpha_kuwata — Function

alpha_kuwata(λ, ε, ε_m, Size)

Principal polarisability components of a spheroidal particle

λ: wavelength

ε: complex dielectric function

ε_m: dielectric function of surrounding medium

Size: SVector with 3 semi-axes of the spheroid

Examples

julia> alpha_kuwata(500, -10+1im, SVector(30, 30, 50), 1.33^2)
3-element SVector{3, ComplexF64} with indices SOneTo(3):
  77076.04648078184 + 26235.664281642235im
  77076.04648078184 + 26235.664281642235im
 -98187.15974124733 + 205835.30299929058im

CoupledDipole.alpha_spheroids — Function

alpha_spheroids(λ, ε, ε_m, Sizes)

Principal polarisability components of N spheroidal particles

λ: wavelength

ε: complex dielectric function

ε_m: dielectric function of surrounding medium

Sizes: Vector of 3-SVectors of particle sizes

Examples

julia> alpha_spheroids(500, -10+1im, 1.33^3, [SVector(30, 30, 50) for i in 1:4])
4-element Vector{SVector{3, ComplexF64}}:
 [83399.81975161123 + 64172.28157743772im, 83399.81975161123 + 64172.28157743772im, -86034.64340475321 + 79773.72029581512im]
 [83399.81975161123 + 64172.28157743772im, 83399.81975161123 + 64172.28157743772im, -86034.64340475321 + 79773.72029581512im]
 [83399.81975161123 + 64172.28157743772im, 83399.81975161123 + 64172.28157743772im, -86034.64340475321 + 79773.72029581512im]
 [83399.81975161123 + 64172.28157743772im, 83399.81975161123 + 64172.28157743772im, -86034.64340475321 + 79773.72029581512im]

Molecular polarisabilities

CoupledDipole.lorentzian — Function

lorentzian(λ::Real, α_k::Real, λ_k::Real, µ_k::Real)

Complex Lorentz function, to describe polarisabilities

λ: wavelength in nm

α_k: oscillator strength in S.I. units

λ_k: oscillator wavelength in nm

µ_k: damping in S.I. units

Examples

julia> round(lorentzian(632.8)*1e39, digits=5)
6.58095 + 1.35961im

CoupledDipole.alpha_bare — Function

alpha_bare(λ::T, α_∞::T, α_k::Array{T}, λ_k::Array{T}, µ_k::Array{T}) where T <: Real

Complex scalar polarisability, as sum of lorentz oscillators

λ: wavelength in nm

α_k: oscillator strength(s) in S.I. units

λ_k: oscillator wavelength(s) in nm

µ_k: damping(s) in S.I. units

Default values mimic the main resonance of Rhodamine 700

Examples

julia> round(alpha_bare(632.8), digits=5)
0.14543 + 0.01222im

CoupledDipole.alpha_embedded — Function

alpha_embedded(α::Complex{T}, medium::T) where T <: Real

Effective point polarisability in medium, rescaled by local field correction

α: bare polarisabilty

medium: refractive index of embedding medium

Default values mimic the main resonance of Rhodamine 700

Examples

julia> round(alpha_embedded(alpha_bare(632.8)), digits=5)
0.12976 + 0.0109im

CoupledDipole.alpha_rescale_molecule — Function

alpha_rescale_molecule(alpha, sizes::Vector{SVector{3}})

Principal polarisability components of a particle, rescaled along each principal axis

α: scalar polarisabilty

sizes: array of 3-vectors to scale along each principal axis