CoupledDipole

Documentation for CoupledDipole.

CoupledDipole.absorption!

CoupledDipole.alpha_bare

CoupledDipole.alpha_embedded

CoupledDipole.alpha_kuwata

CoupledDipole.alpha_rescale_molecule

CoupledDipole.alpha_spheroids

CoupledDipole.cluster_dimer

CoupledDipole.cluster_single

CoupledDipole.cubature_sphere

CoupledDipole.depolarisation_spheroid

CoupledDipole.epsilon_Ag

CoupledDipole.epsilon_Au

CoupledDipole.euler_active

CoupledDipole.euler_passive

CoupledDipole.extinction!

CoupledDipole.incident_field!

CoupledDipole.lorentzian

CoupledDipole.polarisation!

CoupledDipole.propagator_freespace_labframe!

CoupledDipole.quadrature_lgwt

CoupledDipole.scattering!

CoupledDipole.spectrum_dispersion

CoupledDipole.spectrum_oa

CoupledDipole.absorption! — Method

absorption(kn::Real, P::Array{Complex}, E::Array{Complex})

Absorption cross-section for each incident angle

kn: wavenumber in incident medium

P: 3Ndip x Ninc matrix, polarisations for all incidences

E: 3Ndip x Ninc matrix, total field for all incidences

CoupledDipole.alpha_bare — Function

alpha_bare(λ::T, α_∞::T, α_k::Array{T}, λ_k::Array{T}, µ_k::Array{T}) where T <: Real

Complex scalar polarisability, as sum of lorentz oscillators

λ: wavelength in nm

α_k: oscillator strength(s) in S.I. units

λ_k: oscillator wavelength(s) in nm

µ_k: damping(s) in S.I. units

Default values mimic the main resonance of Rhodamine 700

@example alpha_bare(632.8)

CoupledDipole.alpha_embedded — Function

alpha_embedded(α::Complex{T}, medium::T) where T <: Real

Effective point polarisability in medium, rescaled by local field correction

α: bare polarisabilty

medium: refractive index of embedding medium

Default values mimic the main resonance of Rhodamine 700

@example alphaembedded(alphabare(632.8))

CoupledDipole.alpha_kuwata — Method

alpha_kuwata(λ, ε, ε_m, Size)

Principal polarisability components of a spheroidal particle

λ: wavelength

ε: complex dielectric function

ε_m: dielectric function of surrounding medium

Size: SVector with 3 semi-axes of the spheroid

@example alpha_kuwata(500, -10+1im, SVector(30, 30, 50), 1.33^2)

CoupledDipole.alpha_rescale_molecule — Method

alpha_rescale_molecule(alpha, sizes::Vector{SVector{3}})

Principal polarisability components of a particle, rescaled along each principal axis

α: scalar polarisabilty

sizes: array of 3-vectors to scale along each principal axis

@example sizes = [SVector{3}(1.0, 2.0, 3.0) for i in 1:4] alpharescalemolecule(alpha_bare(632.8), sizes)

CoupledDipole.alpha_spheroids — Method

alpha_spheroids(λ, ε, ε_m, Sizes)

Principal polarisability components of N spheroidal particles

λ: wavelength

ε: complex dielectric function

ε_m: dielectric function of surrounding medium

Sizes: Vector of 3-SVectors of particle sizes

@example alpha_spheroids(500, -10+1im, 1.33^3, [SVector(30, 30, 50) for i in 1:4])

CoupledDipole.cluster_dimer — Function

cluster_dimer(d::T, a::T, b::T, c::T, dihedral::T = 0.0, α_1::T = 0.0, α_2::T = 0.0) where T <: Real

Particle cluster consisting of 2 identical particles separated along y

a,b,c: semi-axes along x,y,z

dihedral: angle between both particles seen along the y-axis

material: String referencing the material

type: String, "point" dipole or "particle"

@example cluster_dimer(10,1,2,3)

CoupledDipole.cluster_single — Function

cluster_single(a::T, b::T, c::T, α::T = 0.0, β::T = 0.0, γ::T = 0.0) where T <: Real

Particle cluster consisting of a single particle at the origin

a,b,c: semi-axis along x,y,z

α,β,γ: Euler angles

material: String referencing the material of the particle

type: String, "point" dipole or "particle"

@example cluster_single(1.0,2.0,3.0)

CoupledDipole.cubature_sphere — Function

cubature_sphere(N::Int, method::String)

N-point cubature on the sphere

N: number of nodes

method: cubature method (only 'gl' currently implemented)

Returns a Cubature object containing 2 arrays (N'x3 nodes and N'x1 weights), N'≈N Note: using array instead of tuple for weights because we'll use them in a scalar product in orientation-averaging For nodes there is less of a reason, but it can be convenient to visualise the nodes.

The cubature is normalised by 4π such that a unit integrand approximates 1.

@example cubature_sphere(36*18)

CoupledDipole.depolarisation_spheroid — Method

depolarisation_spheroid(a, b, c)

Depolarisation factor of a spheroid

a: semi-axis along x and y

b: semi-axis along x and y (unused)

c: semi-axis along z

@example depolarisation_spheroid(1, 1, 1.5)

CoupledDipole.epsilon_Ag — Method

epsilon_Ag(λ::Real)

Drude model for the dielectric function of silver in the visible region

λ: wavelength in nm

@example epsilon_Ag(632.8)

CoupledDipole.epsilon_Au — Method

epsilon_Au(λ::Real)

Extended Drude model for the dielectric function of gold in the visible region

λ: wavelength in nm

@example epsilon_Au(632.8)

CoupledDipole.euler_active — Function

euler_active(φ::Real, θ::Real, ψ::Real)

3D rotation matrix

φ: Euler angle (longitude, in [0,2π])

θ: Euler angle (colatitude, in [0,2π])

ψ: Euler angle (rotation around z", in [0,2π])

@example euler_active(π/2,0,0)

CoupledDipole.euler_passive — Function

euler_passive(φ::Real, θ::Real, ψ::Real)

3D rotation matrix

φ: Euler angle (longitude, in [0,2π])

θ: Euler angle (colatitude, in [0,2π])

ψ: Euler angle (rotation around z", in [0,2π])

@example euler_passive(π/2,0,0)

CoupledDipole.extinction! — Method

extinction(kn::Real, P::Array{Complex}, Ein::Array{Complex})

Extinction cross-section for each incident angle

kn: wavenumber in incident medium

P: 3Ndip x Ninc matrix, polarisations for all incidences

Ein: 3Ndip x Ninc matrix, incident field for all incidences

CoupledDipole.incident_field! — Method

incident_field!(Ein,
    Ejones,
    kn, R::Vector{SVector{3}},
    IncidenceRotations)

Incident field at particle positions

Ein: 3N_dip x N_inc matrix, right-hand side of coupled-dipole system

Ejones: tuple of 2 2-Svectors defining 2 orthogonal Jones polarisations

kn: wavenumber in incident medium

R: N_dip-vector of 3-Svectors of particle positions

IncidenceRotations: N_inc-vector of rotation 3-Smatrices

CoupledDipole.lorentzian — Function

lorentzian(λ::T, α_k::T, λ_k::T, µ_k::T) where T <: Real

Complex Lorentz function, to describe polarisabilities

λ: wavelength in nm

α_k: oscillator strength in S.I. units

λ_k: oscillator wavelength in nm

µ_k: damping in S.I. units

@example lorentzian(632.8)

CoupledDipole.polarisation! — Method

polarisation!(P, E, AlphaBlocks)

Self-consistent dipole moments from the electric field, P = αE

P: 3N_dip x N_inc matrix, polarisations for all incidences

E: 3N_dip x N_inc matrix, total field for all incidences

AlphaBlocks: N_dip-vector of 3x3 Smatrices (polarisability tensors in the lab frame)

CoupledDipole.propagator_freespace_labframe! — Method

propagator_freespace_labframe!(A,
    kn, R::Vector{SVector{3}},
    AlphaBlocks::Vector{SMatrix{3,3}})

Interaction matrix

A: 3N_dip x 3N_dip interaction matrix

kn: wavenumber in incident medium

R: N_dip-vector of 3-Svectors of particle positions

AlphaBlocks: N_dip-vector of 3x3 Smatrices (polarisability tensors in the lab frame)

CoupledDipole.quadrature_lgwt — Method

quadrature_lgwt(N::Int, a::Real, b::Real)

N-point Gauss-Legendre quadrature over [a,b] interval

N: number of nodes

a,b: bounds

\[\int_a^b f(x)\,dx=\frac{b-a}{2}\int_{-1}^{1}f\left(\frac{b-a}{2} x + \frac{a+b}{2}\right)\,dx.\]

@example quadrature_lgwt(6,0,3)

CoupledDipole.scattering! — Method

scattering(positions, angles, weights, kn::Real, P::Array{Complex})

Scattering cross-section for each incident angle, obtained by numerical cubature over the full solid angle of scattering directions

positions: vector of cluster particle positions

ScatteringRotations: N_inc-vector of 3x3 rotation Smatrices

weights: N_inc-vector of cubature weights

kn: wavenumber in incident medium

P: 3Ndip x Ninc matrix, polarisations for all incidences

CoupledDipole.spectrum_dispersion — Function

 spectrum_dispersion(cl::Cluster, mat::Material,
                     Incidence, N_sca::Int=36)

Simulating far-field cross-sections for multiple wavelengths and directions of incidence

cl: cluster of particles

mat: dielectric functions

Incidence: N_inc vector of 3-Svectors of incidence Euler angles

N_sca: number of scattering angles for spherical cubature estimate of σ_sca

CoupledDipole.spectrum_oa — Function

 spectrum_oa(cl::Cluster, mat::Material,
            Cubature = "gl", N_inc::Int = 36, N_sca::Int=36)

Orientation-averaged far-field cross-sections for multiple wavelengths

cl: cluster of particles

mat: dielectric functions

Cubature: spherical cubature method

N_inc: number of incident angles for spherical cubature

N_sca: number of scattering angles for spherical cubature estimate of σ_sca