Materials

CoupledDipole.epsilon_Ag — Function

epsilon_Ag(λ::Real)

Drude model for the dielectric function of silver in the visible region

λ: wavelength in nm

Examples

julia> round(epsilon_Ag(632.8), digits=5)
-16.11377 + 0.74871im

CoupledDipole.epsilon_Au — Function

epsilon_Au(λ::Real)

Extended Drude model for the dielectric function of gold in the visible region

λ: wavelength in nm

Examples

julia>  round(epsilon_Au(632.8), digits=5)
-11.40185 + 1.18679im